[viz] upgrade to 0.3.0 + TS
A slew of minor changes done mostly while developing the library.

TS Config largely identical to that in the forecasts-viz library
