+++ viz/README.md
@@ -24,7 +24,13 @@ The GitHub Pages workflow at `.github/workflows/deploy-viz-app.yaml` builds and 
 ### Development
 
 To develop only the forecasts-ncov code whilst continuing to use the library
-from the tarball, `npm run dev` runs [Vite](https://vite.dev) with React Fast Refresh so edits to JSX in `src/main.jsx` (or its imports) update the running app without a full reload.
+from the tarball, `npm run dev` runs [Vite](https://vite.dev) with React Fast Refresh so edits to `src/main.tsx` (or its imports) update the running app without a full reload.
+
+Explicit type-checking is available via:
+
+```sh
+npm run typecheck
+```
 
 To simultaneously develop the viz library alongside this app, ensure the viz components live in the sibling repo
 [nextstrain/forecasts-viz](https://github.com/nextstrain/forecasts-viz),
@@ -47,12 +53,12 @@ that this app ships with:
 
 ## Where things are defined
 
-`./index.html` is the entrypoint. It loads `./src/main.jsx`, which
+`./index.html` is the entrypoint. It loads `./src/main.tsx`, which
 renders the panels and contains the config that controls which panels
 to render. (One day we can hopefully drop React entirely.)
 
 The underlying model JSONs are fetched from S3 via
-`https://data.nextstrain.org/` URLs, as defined in `./src/main.jsx`.
+`https://data.nextstrain.org/` URLs, as defined in `./src/main.tsx`.
 We can add the option to serve local JSONs as needed.
 
 `nextstrain-evofr-viz-*.tgz` is our
