+++ viz/index.html
@@ -19,7 +19,7 @@
       This use is allowable under the <a href="https://www.gisaid.org/registration/terms-of-use/">GISAID Terms of Use</a>.
     </p>
   </main>
-  <script type="module" src="/src/main.jsx"></script>
+  <script type="module" src="/src/main.tsx"></script>
 </body>
 
 </html>
